[SCRUM-4260] Modifications for allowing using last successful data fetched by "scrapers" in case of errors.
From the ticket:
Currently in rdap-ingressd, if one of the two scrapers can't access its
sources, it is not considered anymore for the routing. We need to
allow the last data downloaded from these sources until it can be
access again. Besides we also need to create a health endpoint so we
can allow monitoring the service and check if their data sources are
updated.

In addition we also have implemented a health endpoint for allowing
monitoring the scrapers status.